# docs/openapi.yaml
openapi: 3.1.0

info:
  title: Dice Game Service
  version: 1.0.0
  description: |
    Production-ready dice game backend.

    Every response — success or failure — carries `meta.requestId`, which
    correlates to every log line emitted while handling that request. Quote it
    when reporting a problem.

    Clients should branch on the stable `error.code` string, never on
    `error.message`, which is free to change.
  license:
    name: MIT

servers:
  - url: http://localhost:3000
    description: Local development

tags:
  - name: Games
    description: Create games and play rounds
  - name: Leaderboard
  - name: Health
    description: Infrastructure probes (unversioned, never rate limited)

paths:
  /api/v1/games:
    post:
      tags: [Games]
      summary: Create a game
      operationId: createGame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema: { type: string }
              description: URI of the created game
            x-request-id:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }

    get:
      tags: [Games]
      summary: List games, newest first
      operationId: listGames
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: A page of games
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: object
                    required: [items, pagination]
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Game' }
                      pagination:
                        $ref: '#/components/schemas/Pagination'
                  meta: { $ref: '#/components/schemas/Meta' }
        '400': { $ref: '#/components/responses/BadRequest' }

  /api/v1/games/{gameId}:
    get:
      tags: [Games]
      summary: Fetch a game
      operationId: getGame
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: The game
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GameEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }

  /api/v1/games/{gameId}/rolls:
    post:
      tags: [Games]
      summary: Play a round
      operationId: rollDice
      description: |
        Rolls two dice, scores the round, and appends it to the game.

        The game auto-completes when its final round is played — there is no
        separate "finish" call. Rolling a completed game returns 409.

        Concurrent rolls against the same game are serialised, so no round is
        ever lost.
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '201':
          description: The round played, plus the updated game
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: object
                    required: [round, game]
                    properties:
                      round: { $ref: '#/components/schemas/Round' }
                      game: { $ref: '#/components/schemas/Game' }
                  meta: { $ref: '#/components/schemas/Meta' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: |
            The game has already played all its rounds
            (`GAME_ALREADY_COMPLETED`), or the write lost an optimistic
            concurrency check (`CONCURRENCY_CONFLICT` — re-read and retry).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '429': { $ref: '#/components/responses/TooManyRequests' }

  /api/v1/leaderboard:
    get:
      tags: [Leaderboard]
      summary: Top completed games by score
      operationId: getLeaderboard
      description: |
        Only completed games appear. Ties are broken by earliest completion, so
        ordering is stable across calls.
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
      responses:
        '200':
          description: Ranked entries
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/LeaderboardEntry' }
                  meta: { $ref: '#/components/schemas/Meta' }

  /healthz:
    get:
      tags: [Health]
      summary: Liveness probe
      operationId: live
      description: |
        Answers "is this process wedged?" and depends on nothing downstream —
        a dependency outage must never cause a restart storm.
      responses:
        '200':
          description: The process is alive
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, examples: ['ok'] }
                  service: { type: string }
                  version: { type: string }
                  uptimeSeconds: { type: integer }
                  timestamp: { type: string, format: date-time }

  /readyz:
    get:
      tags: [Health]
      summary: Readiness probe
      operationId: ready
      description: |
        Answers "should traffic route here right now?". Returns 503 as soon as
        SIGTERM is received, so the load balancer drains this instance before
        the server stops accepting connections.
      responses:
        '200':
          description: Ready to serve
        '503':
          description: Draining — do not route new traffic

components:
  parameters:
    GameId:
      name: gameId
      in: path
      required: true
      schema: { type: string, format: uuid }

  schemas:
    CreateGameRequest:
      type: object
      required: [playerName]
      additionalProperties: false
      properties:
        playerName:
          type: string
          minLength: 1
          maxLength: 64
          description: Letters, numbers, spaces, dots, underscores and hyphens.
          examples: ['Ada Lovelace']
        rounds:
          type: integer
          minimum: 1
          maximum: 100
          description: |
            Defaults to `GAME_DEFAULT_ROUNDS` (5). Values above the configured
            `GAME_MAX_ROUNDS` (20) are well-formed but domain-invalid → 422.
          examples: [3]

    Game:
      type: object
      required:
        [
          id,
          playerName,
          status,
          totalRounds,
          roundsPlayed,
          roundsRemaining,
          totalScore,
          rounds,
          createdAt,
          completedAt,
        ]
      properties:
        id: { type: string, format: uuid }
        playerName: { type: string }
        status: { type: string, enum: [IN_PROGRESS, COMPLETED] }
        totalRounds: { type: integer }
        roundsPlayed: { type: integer }
        roundsRemaining: { type: integer }
        totalScore: { type: integer }
        rounds:
          type: array
          items: { $ref: '#/components/schemas/Round' }
        createdAt: { type: string, format: date-time }
        completedAt:
          type: [string, 'null']
          format: date-time

    Round:
      type: object
      required: [index, dice, pips, outcome, points, scoreAfter, rolledAt]
      properties:
        index:
          type: integer
          description: 1-based position within the game.
        dice:
          type: array
          minItems: 2
          maxItems: 2
          items: { type: integer, minimum: 1, maximum: 6 }
        pips:
          type: integer
          description: Sum of both dice.
        outcome:
          type: string
          enum: [SNAKE_EYES, DOUBLES, LUCKY_SEVEN, STANDARD]
          description: |
            Evaluated in strict precedence order:
              SNAKE_EYES  (1,1) ....... 0 points, overrides the doubles bonus
              DOUBLES     (n,n) ....... pips x 2
              LUCKY_SEVEN (pips = 7) .. pips + 10
              STANDARD ................ pips
        points: { type: integer, minimum: 0, maximum: 24 }
        scoreAfter:
          type: integer
          description: Running total after this round.
        rolledAt: { type: string, format: date-time }

    LeaderboardEntry:
      type: object
      required: [rank, gameId, playerName, totalScore, totalRounds, completedAt]
      properties:
        rank: { type: integer, minimum: 1 }
        gameId: { type: string, format: uuid }
        playerName: { type: string }
        totalScore: { type: integer }
        totalRounds: { type: integer }
        completedAt:
          type: [string, 'null']
          format: date-time

    Pagination:
      type: object
      required: [total, limit, offset, hasMore]
      properties:
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
        hasMore: { type: boolean }

    Meta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId:
          type: string
          description: Correlates to every log line for this request.
        timestamp: { type: string, format: date-time }

    GameEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/Game' }
        meta: { $ref: '#/components/schemas/Meta' }

    ErrorEnvelope:
      type: object
      required: [error, meta]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable identifier. Branch on this, never on `message`.
              enum:
                - VALIDATION_ERROR
                - MALFORMED_REQUEST_BODY
                - PAYLOAD_TOO_LARGE
                - GAME_NOT_FOUND
                - ROUTE_NOT_FOUND
                - GAME_ALREADY_COMPLETED
                - CONCURRENCY_CONFLICT
                - INVALID_ROUND_COUNT
                - RATE_LIMIT_EXCEEDED
                - INTERNAL_SERVER_ERROR
            message: { type: string }
            details:
              description: Machine-readable context. Never contains secrets.
        meta: { $ref: '#/components/schemas/Meta' }

  responses:
    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    UnprocessableEntity:
      description: Well-formed but violates a domain rule
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }